Move the type predicates to the internal folder
