+++ src/rlang/vec.h
@@ -101,17 +101,6 @@ void r_list_poke(sexp* list, r_ssize i, sexp* elt) {
 }
 
 
-bool r_is_vector(sexp* x, r_ssize n);
-bool r_is_atomic(sexp* x, r_ssize n);
-
-bool r_is_finite(sexp* x);
-bool r_is_logical(sexp* x, r_ssize n);
-bool r_is_integerish(sexp* x, r_ssize n, int finite);
-bool r_is_integer(sexp* x, r_ssize n, int finite);
-bool r_is_double(sexp* x, r_ssize n, int finite);
-bool r_is_character(sexp* x, r_ssize n);
-bool r_is_raw(sexp* x, r_ssize n);
-
 static inline
 bool r_is_int(SEXP x) {
   return r_typeof(x) == r_type_integer &&
